@@ -38,7 +38,6 @@ classifiers =
   Programming Language :: Cython
   Programming Language :: Python
   Programming Language :: Python :: 3
-  Programming Language :: Python :: 3.8
   Programming Language :: Python :: 3.9
   Programming Language :: Python :: 3.10
   Programming Language :: Python :: 3.11
@@ -53,7 +52,7 @@ keywords =
   propcache
 
 [options]
-python_requires = >=3.8
+python_requires = >=3.9
 # Ref:
 # https://setuptools.pypa.io/en/latest/userguide/declarative_config.html#using-a-src-layout
 # (`src/` layout)
